poc_requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
packaging==24.2
pillow==11.1.0
pytesseract==0.3.13
requests==2.32.3
urllib3==2.4.0
